Spigot-Bukkit-1.8.x-AGift
I have written this plugin for my skyblock server to give daily gifts
for the players. You can customize the items in Gifts.yml, also you can
change things like display name with colors which you like and you can
change the amount of the items, the selection is random.